pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["BeemenSameh <41084077+BeemenSameh@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14"
django-environ = "^0.10"
psycopg = "^3.1"
gunicorn = "^20.1"
uWSGI = "^2.0"
django-extensions = "^3.2"

[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21"
docutils = "^0.19"
isort = "^5.12"
pytest-django = "^4.5"
ipython = "^8.12"
coverage = "^7.2"
autopep8 = "^2.0"
django-debug-toolbar = "^4.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
